fix: Working semver
The semver of `standard-version` seems to not work as expected when the
version is below v1.0.0. Now that I manually bumped the major version,
it seems to be working fine now. This fix commit should bump the patch
version to v1.1.1.

One more (minor) thing to fix would be the generated changelogs of the
self-hosted runner, which isn't attaching the links to the header
properly. Not a big deal, but would be nice to have working.